refactor: move the field-value reading into Documents_Meta_Handler
Seven helpers sat in the CPT while being shared between rendering the editor
and saving it: reading a post's structured values, decoding a repeater's
stored rows, normalising an item schema, and humanising a slug into a label.
They belong with parse_structured_content() and get_meta_fields_for_post(),
which already live in Documents_Meta_Handler and do the same kind of work.

Moving them is what makes the class split possible: it is the only set of
methods both the metabox renderer and the save pipeline reach into, so
leaving them behind would have forced one of those extractions to call back
into the CPT.

fix_unescaped_unicode_sequences() turns out to be a third duplicate - the
CPT's copy is byte-identical to the handler's - so the CPT's is dropped
rather than moved.

Two things stay put deliberately. ARRAY_FIELD_MAX_ITEMS and
decode_array_field_value() are asserted by name against Documentate_Documents
in the test suite, so both keep that home: the constant now takes its value
from the handler and the method delegates, which leaves the published surface
untouched while the implementation moves.

Eleven tests reached these through ReflectionClass on a private method. They
now call the public static methods directly, which is what they wanted to say
in the first place.

Markup verified identical, by the same dump-and-diff used for the previous
two commits. Class complexity drops 495 to 446.
